cut down on redundant inclusions of <sys/types.h>, include it in proto.h per DB's patch, and make a few more minor cosmetic cleanups
git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@1857 35c25a1d-7b9e-4130-9fde-d3aeb78583b8

+++ src/winio.c
@@ -2183,7 +2183,6 @@ void reset_cursor(void)
     current_y = current->lineno - edittop->lineno;
     if (current_y < editwinrows) {
 	size_t x = xplustabs();
-
 	wmove(edit, current_y, x - get_page_start(x));
      }
 }
@@ -3212,7 +3211,7 @@ void dump_buffer_reverse(void)
     const filestruct *fileptr = filebot;
 
     while (fileptr != NULL) {
-	fprintf(stderr, "(%d) %s\n", fileptr->lineno, fileptr->data);
+	fprintf(stderr, "(%lu) %s\n", (unsigned long)fileptr->lineno, fileptr->data);
 	fileptr = fileptr->prev;
     }
 }
